# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: CHEMIA Chemistry Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /formulas:
    get:
      operationId: listFormulas
      tags: [formulas]
      summary: List saved formulas
      responses:
        "200":
          description: Saved formulas
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SavedFormula"
    post:
      operationId: saveFormula
      tags: [formulas]
      summary: Save a formula
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SavedFormulaInput"
      responses:
        "201":
          description: Saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SavedFormula"
  /formulas/{id}:
    delete:
      operationId: deleteFormula
      tags: [formulas]
      summary: Delete a saved formula
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    SavedFormula:
      type: object
      properties:
        id:
          type: integer
        formula:
          type: string
        name:
          type: string
        type:
          type: string
        molecularMass:
          type: ["number", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - formula
        - name
        - type
        - createdAt
    SavedFormulaInput:
      type: object
      properties:
        formula:
          type: string
        name:
          type: string
        type:
          type: string
        molecularMass:
          type: ["number", "null"]
      required:
        - formula
        - name
        - type
...