# lib/api-zod/src/generated/types/savedFormulaInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * CHEMIA Chemistry Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface SavedFormulaInput {
  formula: string;
  name: string;
  type: string;
  /** @nullable */
  molecularMass?: number | null;
}